Add textScaleFactor to Linkify Widget
textScaleFactor is an important parameter to preverse
accessibility in text Widgets, since we need to scale
the font size depending on what the user has configured
in their operating system.

If the developer does not provide any value, the system
scale factor will be used.

If the developer provides a specific value, it will override
the defaults.